#2067 - after discussion
